Update textvation.c
Added welcome screen. You can now use arrow keys to navigate the program and the pgUp and pdDn keys take you to the bottom and top of the page. The Home and End keys are supposed to take you to the very left and right of a page but we can't really print anything yet which I will have to go through and try to figure out