Update textvation.c
I have successfully added some level of syntax highlighting.
Numbers are now red and when searching text, it is highlighted in blue. Comments are cyan and so on. I will update the colors down the line to something more aesthetic (Maybe Catpuccin). 

This will currently only support syntax highlighting for C, C++ and .h files.

We're basically in the end stretch now. Once this is done, I will be debugging all the issues, occasionally adding new features but we're almost at v1.0